Доработал по замечаниям
