+++ js/filters.js
@@ -1,7 +1,7 @@
 const SHUFFLED_LIMIT = 10;
 
-const getShuffledPhotos = (photos) => photos.slice().sort(() => Math.random() - 0.5).slice(0, SHUFFLED_LIMIT);
-const getDiscussedPhotos = (photos) => photos.slice().sort((a, b) => b.comments.length - a.comments.length);
+const getShuffledPhotos = (pictures) => pictures.slice().sort(() => Math.random() - 0.5).slice(0, SHUFFLED_LIMIT);
+const getDiscussedPhotos = (pictures) => pictures.slice().sort((a, b) => b.comments.length - a.comments.length);
 
 export {
   getShuffledPhotos,
