+++ js/util/dom.js
@@ -1,0 +1,5 @@
+const removeElement = (element) => {
+  element.remove();
+};
+
+export {removeElement};
